requirements.txt
tensorflow
# opencv-python
pandas
pillow
flask